add rtc test
